adds Request::mapTo
