--- src/Lemon/DataMarshaller/README.md
@@ -1,3 +1,0 @@
-# Data Marshaller
-
-This class converts data from asociative array to class.
